Fixed wrong event grab

+++ interactions/buttons/games/deleteEvent.js
@@ -13,7 +13,7 @@ module.exports = {
     await interaction.deferReply({ ephemeral: true })
     const events = interaction.guild.scheduledEvents.cache
     for (let event of events.values()) {
-      if (event.description.includes(interaction.user.toString())) {
+      if (event.description.includes(interaction.user.toString()) && interaction.message.embeds[0].author.name.includes(interaction.user.tag)) {
         try {
           event.delete()
         } catch {
